add role management and caching

+++ VollyV2/Controllers/Mvc/OpportunitiesController.cs
@@ -1,38 +1,75 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.ChangeTracking;
+using Microsoft.EntityFrameworkCore.Query;
 using VollyV2.Data;
 using VollyV2.Data.Volly;
+using VollyV2.Models;
 using VollyV2.Models.Volly;
+using VollyV2.Services;
 
 namespace VollyV2.Controllers.Mvc
 {
-    [Authorize(Roles = "Admin")]
+    [Authorize(Roles = "Admin, User")]
     public class OpportunitiesController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public OpportunitiesController(ApplicationDbContext context)
+        public OpportunitiesController(ApplicationDbContext context,
+            UserManager<ApplicationUser> userManager,
+            IAuthorizationService authorizationService)
         {
             _context = context;
+            _userManager = userManager;
+            _authorizationService = authorizationService;
         }
 
         // GET: Opportunities
         public async Task<IActionResult> Index()
         {
-            List<Opportunity> opportunities = await _context.Opportunities
+            ApplicationUser user = await GetCurrentUser();
+            bool isAdmin = await IsAdmin(user);
+            IIncludableQueryable<Opportunity, Organization> opportunitiesQueryable = _context.Opportunities
                 .Include(o => o.Category)
-                .Include(o => o.Organization)
-                .ToListAsync();
+                .Include(o => o.Organization);
+
+            List<Opportunity> opportunities;
+            if (isAdmin)
+            {
+                opportunities = await opportunitiesQueryable
+                    .ToListAsync();
+            }
+            else
+            {
+                opportunities = await opportunitiesQueryable
+                    .Where(o => o.CreatedByUserId == user.Id)
+                    .ToListAsync();
+            }
 
             opportunities.ForEach(a => a.DateTime = TimeZoneInfo.ConvertTimeFromUtc(a.DateTime, VollyConstants.TimeZoneInfo));
-        
+
             return View(opportunities);
+        }
+
+        private async Task<bool> IsAdmin(ApplicationUser user)
+        {
+            return await _userManager.IsInRoleAsync(user, "Admin");
+        }
+
+        private async Task<ApplicationUser> GetCurrentUser()
+        {
+            ApplicationUser user = await _userManager.GetUserAsync(User);
+            return user;
         }
 
         // GET: Opportunities/Details/5
@@ -52,7 +89,16 @@ namespace VollyV2.Controllers.Mvc
             {
                 return NotFound();
             }
-            opportunity.DateTime = TimeZoneInfo.ConvertTimeFromUtc(opportunity.DateTime, VollyConstants.TimeZoneInfo);
+
+            AuthorizationResult authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, opportunity, new OpportunityCreatorRequirement());
+
+            if (!authorizationResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
+            opportunity.DateTime =
+                TimeZoneInfo.ConvertTimeFromUtc(opportunity.DateTime, VollyConstants.TimeZoneInfo);
 
             return View(opportunity);
         }
@@ -82,6 +128,7 @@ namespace VollyV2.Controllers.Mvc
             if (ModelState.IsValid)
             {
                 Opportunity opportunity = model.GetOpportunity(_context);
+                opportunity.CreatedByUser = await GetCurrentUser();
                 _context.Add(opportunity);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -106,6 +153,14 @@ namespace VollyV2.Controllers.Mvc
                 return NotFound();
             }
 
+            AuthorizationResult authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, opportunity, new OpportunityCreatorRequirement());
+
+            if (!authorizationResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
+
             return View(OpportunityModel.FromOpportunity(_context, opportunity));
         }
 
@@ -123,23 +178,26 @@ namespace VollyV2.Controllers.Mvc
 
             if (ModelState.IsValid)
             {
-                Opportunity opportunity = new Opportunity();
                 try
                 {
-                    opportunity = opportunityModel.GetOpportunity(_context);
+                    Opportunity opportunity = opportunityModel.GetOpportunity(_context);
+                    AuthorizationResult authorizationResult = await _authorizationService
+                        .AuthorizeAsync(User, opportunity, new OpportunityCreatorRequirement());
+
+                    if (!authorizationResult.Succeeded)
+                    {
+                        return new ForbidResult();
+                    }
                     _context.Update(opportunity);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!OpportunityExists(opportunity.Id))
+                    if (!OpportunityExists(opportunityModel.Id))
                     {
                         return NotFound();
                     }
-                    else
-                    {
-                        throw;
-                    }
+                    throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -158,9 +216,17 @@ namespace VollyV2.Controllers.Mvc
                 .Include(o => o.Category)
                 .Include(o => o.Organization)
                 .SingleOrDefaultAsync(m => m.Id == id);
+
             if (opportunity == null)
             {
                 return NotFound();
+            }
+            AuthorizationResult authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, opportunity, new OpportunityCreatorRequirement());
+
+            if (!authorizationResult.Succeeded)
+            {
+                return new ForbidResult();
             }
             opportunity.DateTime = TimeZoneInfo.ConvertTimeFromUtc(opportunity.DateTime, VollyConstants.TimeZoneInfo);
 
@@ -173,6 +239,13 @@ namespace VollyV2.Controllers.Mvc
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var opportunity = await _context.Opportunities.SingleOrDefaultAsync(m => m.Id == id);
+            AuthorizationResult authorizationResult = await _authorizationService
+                .AuthorizeAsync(User, opportunity, new OpportunityCreatorRequirement());
+
+            if (!authorizationResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
             _context.Opportunities.Remove(opportunity);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
